pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamlineagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "facebook-business>=23.0.1",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "mcp>=1.12.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-facebook-api>=0.20.1",
    "uvicorn>=0.35.0",
    "requests>=2.31.0",
    "openai>=1.99.9",
    "flask>=3.0.0",
    "bcrypt>=4.0.1",
    "PyPDF2>=3.0.0",
    "numpy>=1.24.0",
    "supabase>=2.0.0",
    "anyio>=4.0.0",
    "jsonschema>=4.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "utils*"]
exclude = ["frontend*", "pdfs*", "test*", "*.test*"]
